requirements.txt
gym==0.22
torch
torchvision
numpy
matplotlib
pygame

wandb
tensorboard

hydra-core
omegaconf
